make Fight feature work

+++ internal/api/services/equipment_item_take_on.go
@@ -102,7 +102,7 @@ func (s *EquipmentItemTakeOnService) TakeOnEquipmentItem(ctx context.Context, us
 		return err
 	}
 
-	if uint(user.Level) < item.RequiredLevel {
+	if user.Level < item.RequiredLevel {
 		return ErrInsufficientLevel
 	}
