+++ internal/domain/bot.go
@@ -5,8 +5,8 @@ type Bot struct {
 	Name    string `db:"name"`
 	Slug    string `db:"slug"`
 	Avatar  string `db:"avatar"`
-	Attack  uint16 `db:"attack"`
-	Defense uint16 `db:"defense"`
-	Hp      uint16 `db:"hp"`
-	Level   uint8  `db:"level"`
+	Attack  uint `db:"attack"`
+	Defense uint `db:"defense"`
+	Hp      uint `db:"hp"`
+	Level   uint `db:"level"`
 }
